Move little endian streams to io package
